requirements.txt
opencv-python
numpy
pytesseract
ftfy
spacy
pillow

